app/requirements.txt
pyyaml==5.3
azure.iot.device==2.3.0